requirements.txt
Flask==2.3.0
Flask-Cors==3.0.10
google-generativeai==0.8.1